requirements.txt
fastapi
uvicorn[standard]
playwright
apscheduler
python-dotenv
python-dateutil
jinja2
aiofiles